Печать дипломов

+++ ContestManager/Core/Contests/Result.cs
@@ -7,8 +7,8 @@ namespace Core.Contests
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string SchoolWithCity { get; set; }
-        public int[] Results { get; set; }
-        public int Sum { get; set; }
+        public double[] Results { get; set; }
+        public double Sum { get; set; }
         public string Place { get; set; }
     }
 }
